add secretmanager post-processors
