+++ spring-cloud-gcp-autoconfigure/src/main/resources/META-INF/spring.factories
@@ -31,4 +31,5 @@ com.google.cloud.spring.autoconfigure.config.GcpConfigBootstrapConfiguration,\
 com.google.cloud.spring.autoconfigure.secretmanager.GcpSecretManagerBootstrapConfiguration
 
 org.springframework.boot.env.EnvironmentPostProcessor=\
-com.google.cloud.spring.autoconfigure.sql.CloudSqlEnvironmentPostProcessor
+com.google.cloud.spring.autoconfigure.sql.CloudSqlEnvironmentPostProcessor,\
+com.google.cloud.spring.autoconfigure.secretmanager.GcpSecretManagerEnvironmentPostProcessor
